refactor: router 관심사별 분리 및 폴더 위치이동
