+++ src/app/routes.js
@@ -1,0 +1,6 @@
+import { Home, Mypage } from "../pages";
+
+export const routes = {
+  "/": Home,
+  "/mypage": Mypage,
+};
